apps/geocore-uut/backend/requirements/development.txt
amqp==2.5.2
aniso8601==7.0.0
APScheduler==3.6.3
asgiref==3.2.3
attrs==19.3.0
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
commandbus==1.0.1
coverage==4.5.1
Django==3.0.4
django-appconf==1.0.4
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filter==1.1.0
django-tilestache==2.0.9
djangorestframework==3.11.0
djangorestframework-filters==0.11.1
djangorestframework-gis==0.15
docutils==0.16
Fiona==1.8.13.post1
geopandas==0.7.0
graphene==2.1.8
graphene-django==2.9.1
graphql-core==2.3.1
graphql-relay==2.0.1
idna==2.9
importlib-metadata==1.6.0
jsonfield==3.1.0
kombu==4.6.8
model-mommy==1.5.1
ModestMaps==1.4.7
munch==2.5.0
numpy==1.18.3
pandas==1.0.3
Pillow==7.1.2
promise==2.3
psycopg2-binary==2.8.4
pylru==1.2.0
pyproj==2.6.0
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
Rx==1.6.1
Shapely==1.7.0
simplejson==3.17.0
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
TileStache==1.51.14
tzlocal==2.0.0
urllib3==1.25.8
vine==1.3.0
Werkzeug==1.0.1
zipp==3.1.0
